Gotov projekat, sve radi.

+++ src/main/java/se201/projekat/main/Main.java
@@ -19,49 +19,37 @@ public class Main extends Application {
 
     public void start(Stage primaryStage) throws Exception {
 
+        // TEST PODACI
+        if(true){
+            // Prvo ocisti celu bazu
+            AddressDao addressDao = DaoFactory.create(AddressDao.class);
+            addressDao.deleteAll();
+            GroupDao groupDao = DaoFactory.create(GroupDao.class);
+            groupDao.deleteAll();
+            ContactDao contactDao = DaoFactory.create(ContactDao.class);
+            contactDao.deleteAll();
+            DaoFactory.create(PersonDao.class).deleteAll();
+            DaoFactory.create(GroupDao.class).deleteAll();
 
-       /* AddressDao addressDao = DaoFactory.create(AddressDao.class);
-        addressDao.deleteAll(); // ocisti celu bazu
+            Address adr1 = new Address("Serbia", "Knjazevac", "Ulica 1", "123");
+            Address adr2 = new Address("Serbia", "Beograd", "Ulica 2", "3/C");
+            Person person1 = new Person("Filip","Dimitrijevic", Gender.MALE);
+            Person person2 = new Person("Milica","Aleksic", Gender.FEMALE);
+            Contact c1 = new Contact(person1, adr1, "filipd@gmail.com", "053-234-123");
+            Contact c2 = new Contact(person2, adr2, "milica.aleksic123@hotmail.com", "123-123-123");
 
-        DaoFactory.create(PersonDao.class).deleteAll();
-        DaoFactory.create(GroupDao.class).deleteAll();
+            contactDao.insert(c1);
+            contactDao.insert(c2);
 
-        ContactDao contactDao = DaoFactory.create(ContactDao.class);
-        contactDao.deleteAll();
+            groupDao.insert(new Group("Prijatelji"));
+            groupDao.insert(new Group("Posao"));
+        }
 
-
-
-        Address adr = new Address("Serbia", "Nis", "asd", "123");
-        Address adr1 = new Address("Serbia", "BG", "wwwww", "3A/4");
-
-        addressDao.insert(adr);
-        addressDao.insert(adr1);
-
-        adr.setCity("Pirot");
-        addressDao.update(adr);
-        System.out.println(addressDao.getAll());
-        addressDao.delete(adr1);
-        System.out.println(addressDao.getAll());
-        System.out.println("deleted: " + adr1);
-
-        Person person1 = new Person("Filip","Dimitrijevic", Gender.MALE);
-        Person person2 = new Person("Milica","Milivojevic", Gender.FEMALE);
-
-        Contact c1 = new Contact(person1, adr, "filipd@gmail.com", "053-234-123");
-        Contact c2 = new Contact(person2, adr1, "milicam@gmail.com", "123-123-123");
-
-        contactDao.insert(c1);
-        contactDao.insert(c2);
-        System.out.println(contactDao.getAll());
-*/
         Parent root = FXMLLoader.load(getClass().getResource("../MainView.fxml"));
 
         Scene scene = new Scene(root);
         primaryStage.setScene(scene);
         primaryStage.show();
-        primaryStage.setOnCloseRequest(event -> {
-            System.out.println("Closing boi!");
-        });
     }
 
     public static void main(String[] args) {
